people contract 2 实现

+++ src/main/java/com/wangzhixuan/controller/PeopleContract2SalaryController.java
@@ -58,7 +58,7 @@ public class PeopleContract2SalaryController extends BaseController {
 	}
 
 	@RequestMapping(value = "/salaryListPage", method = RequestMethod.GET)
-	public String salaryListPage(Long id, Model model) {
+	public String salaryListPage(Integer id, Model model) {
 
 		PeopleContract people = peopleContract2Service.findPeopleContractById(id);
 		logger.info(JSON.toJSONString(people));
